requirements.txt
# Core
pydantic>=2.0.0
pydantic-settings>=2.0.0
typing-extensions>=4.8.0

# HTTP & API
httpx>=0.25.0
requests>=2.31.0
aiohttp>=3.9.0

# Evaluation metrics
sacrebleu>=2.3.0
meeteval>=0.2.0
editdistance>=0.6.2
jiwer>=3.0.0

# Data processing
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0

# Audio processing
soundfile>=0.12.0
librosa>=0.10.0

# Model download
huggingface-hub>=0.20.0
modelscope>=1.10.0

# Logging & CLI
rich>=13.0.0
structlog>=23.0.0
typer>=0.9.0

# Database
tinydb>=4.8.0

# MCP
mcp>=1.0.0

# Config
pyyaml>=6.0.0
